Integration tests for rust libraries at this stage only include a few transactions within a single files.
This PR will:
- Separate into multiple files (1 file / transaction) for readability and test efficiency.
- Add tests for remaining transactions (using xrpl.js as benchmark)
- Use Docker container for integration tests to develop unreleased features
- Add test coverage checks

+++ src/utils/txn_parser/get_final_balances.rs
@@ -26,7 +26,7 @@ fn compute_final_balance(node: &NormalizedNode) -> XRPLUtilsResult<Option<BigDec
             value = get_value(&balance.clone().into())?;
         }
     }
-    if value == BigDecimal::from(0) {
+    if value == 0 {
         return Ok(None);
     }
     Ok(Some(value))
